Migrated from a card component to a character one
